refactor(collection-manage): добавлен index.ts для реэкспорта компонентов
