+++ frontend/src/features/collection-manage/index.ts
@@ -1,0 +1,3 @@
+export { ActionsMenu } from "./ui/ActionsMenu";
+export { AddCollection } from "./ui/AddCollection";
+export { RenameCollection } from "./ui/RenameCollection";
